Epsilon Badges uses ENSL team ID for badge name
Previously, Epsilon Badges looked for badge image files that match the
ENSL team name. This is prone to change at the whim of players who discover
the latest meme. This commit changes Epsilon Badges to look for badge
image files that match "ENSLX", where X is the ENSL team ID. The team
ID is unique and immutable.