[RRQueue] Rework player slot overflow check
The CheckConnectionAllowed event gets consumed by the reserved slot plugin. This causes the rrqueue plugin to wrongly detect player slot overflows when there isn't one.
So we will only do a simple check after each client connected to ensure that the amount of players is never higher than the max.